Fix some PHP 8.5 depreciations
FossilOrigin-Name: 53d25a5318a1fec5d610d8d90e4b34e872a52964

+++ src/lib/KD2/Karto/Point.php
@@ -145,7 +145,7 @@ class Point implements ArrayAccess
 			throw new \InvalidArgumentException('Invalid latitude (must be between -180 and 180)');
 		}
 
-		$this->$key = (double) $value;
+		$this->$key = (float) $value;
 	}
 
 	/**
